+++ libs/libxx/Makefile
@@ -19,14 +19,6 @@
 
 include $(TOPDIR)/Make.defs
 
-CXXSRCS = libxx_cxa_atexit.cxx
-
-ifeq ($(CONFIG_LIBSUPCXX),y)
-CXXSRCS += libxx_impure.cxx
-else
-CXXSRCS += libxx_eabi_atexit.cxx
-endif
-
 # Include the uClibc++ Make.defs file if selected.  If it is included,
 # the uClibc++/Make.defs file will add its files to the source file list,
 # add its DEPPATH info, and will add the appropriate paths to the VPATH
@@ -42,12 +34,20 @@ include uClibc++.defs
 else ifeq ($(CONFIG_LIBCXX),y)
 include libcxx.defs
 else
-include cxx.defs
+include libcxxmini.defs
 endif
 
 ifeq ($(CONFIG_LIBCXXABI),y)
 include libcxxabi.defs
 endif
+
+ifeq ($(CONFIG_LIBSUPCXX),y)
+CXXSRCS += libxx_impure.cxx
+else
+CXXSRCS += libxx_eabi_atexit.cxx
+endif
+
+CXXSRCS += libxx_cxa_atexit.cxx
 
 # Object Files
