+++ libs/libxx/libcxxmini/libxx_cxa_guard.cxx
@@ -1,5 +1,5 @@
 //***************************************************************************
-// libs/libxx/libxx_cxa_guard.cxx
+// libs/libxx/libcxxmini/libxx_cxa_guard.cxx
 //
 //   Copyright (C) 2015 Omni Hoverboards Inc. All rights reserved.
 //   Author: Paul Alexander Patience <paul-a.patience@polymtl.ca>
