+++ libs/libxx/libcxxmini/libxx_delete.cxx
@@ -1,5 +1,5 @@
 //***************************************************************************
-// libs/libxx/libxx_deletea_sized.cxx
+// libs/libxx/libcxxmini/libxx_delete.cxx
 //
 // Licensed to the Apache Software Foundation (ASF) under one or more
 // contributor license agreements.  See the NOTICE file distributed with
@@ -22,25 +22,18 @@
 //***************************************************************************
 
 #include <nuttx/config.h>
-#include <nuttx/compiler.h>
 
-#include <cstddef>
-
-#include "libxx.hxx"
-
-#ifdef CONFIG_HAVE_CXX14
+#include <nuttx/lib/lib.h>
 
 //***************************************************************************
 // Operators
 //***************************************************************************
 
 //***************************************************************************
-// Name: delete[]
+// Name: delete
 //***************************************************************************
 
-void operator delete[](FAR void *ptr, std::size_t size)
+void operator delete(FAR void *ptr) throw()
 {
   lib_free(ptr);
 }
-
-#endif /* CONFIG_HAVE_CXX14 */
