+++ libs/libxx/libcxxmini/libxx_delete_sized.cxx
@@ -1,5 +1,5 @@
 //***************************************************************************
-// libs/libxx/libxx_deletea.cxx
+// libs/libxx/libcxxmini/libxx_delete_sized.cxx
 //
 // Licensed to the Apache Software Foundation (ASF) under one or more
 // contributor license agreements.  See the NOTICE file distributed with
@@ -21,19 +21,35 @@
 // Included Files
 //***************************************************************************
 
-#include <nuttx/config.h>
+#include <nuttx/compiler.h>
 
-#include "libxx.hxx"
+#include <cstddef>
+
+#include <nuttx/lib/lib.h>
+
+#ifdef CONFIG_HAVE_CXX14
 
 //***************************************************************************
 // Operators
 //***************************************************************************
 
 //***************************************************************************
-// Name: delete[]
+// Name: delete
+//
+// NOTE:
+//   This should take a type of size_t.  But size_t has an unknown underlying
+//   type.  In the nuttx sys/types.h header file, size_t is typed as uint32_t
+//   (which is determined by architecture-specific logic).  But the C++
+//   compiler may believe that size_t is of a different type resulting in
+//   compilation errors in the operator.  Using the underlying integer type
+//   instead of size_t seems to resolve the compilation issues. Need to
+//   REVISIT this.
+//
 //***************************************************************************
 
-void operator delete[](FAR void *ptr) throw()
+void operator delete(FAR void *ptr, std::size_t size)
 {
   lib_free(ptr);
 }
+
+#endif /* CONFIG_HAVE_CXX14 */
