+++ libs/libxx/libxx_cxa_atexit.cxx
@@ -24,6 +24,7 @@
 #include <nuttx/config.h>
 
 #include <cassert>
+#include <nuttx/lib/lib.h>
 
 #include "libxx.hxx"
